Fix Build Error

+++ Classes/NJOutputMouseScroll.m
@@ -36,7 +36,7 @@
 - (int)wheel:(int)n {
     int amount =  abs(_direction) == n ? _direction / n : 0;
     if (self.smooth)
-        amount *= _speed * self.magnitude;
+        amount *= (int) (_speed * self.magnitude);
     return amount;
 }
